admin-panel/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0

